updated task model and made tasks user specific

+++ api/views.py
@@ -1,7 +1,12 @@
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
+from rest_framework.exceptions import NotAuthenticated, NotFound
 from .serializers import TaskSerializer
 from .models import Task
+import jwt
+import os
+
+JWT_SECRET = os.environ.get('JWT_SECRET')
 
 # Create your views here.
 
@@ -25,8 +30,18 @@ def apiOverView(request):
 @api_view(['GET'])
 def taskList(request):
 
+    token = request.COOKIES.get('jwt')
+
+    if not token:
+        raise NotAuthenticated('Unauthenticated')
+
+    try:
+        payload = jwt.decode(token, JWT_SECRET, ['HS256'])
+    except jwt.ExpiredSignatureError:
+        raise NotAuthenticated('Unauthenticated')
+
     # Getting all tasks from DB
-    tasks = Task.objects.all()
+    tasks = Task.objects.filter(user=payload['id'])
 
     # Serializing the database data as API response
     # many=True should be kept when we expect multiple data from model
@@ -36,7 +51,18 @@ def taskList(request):
 
 @api_view(["GET"])
 def taskDetail(request, pk):
-    task = Task.objects.get(id=pk)
+
+    token = request.COOKIES.get('jwt')
+
+    if not token:
+        raise NotAuthenticated('Unauthenticated')
+
+    try:
+        payload = jwt.decode(token, JWT_SECRET, ['HS256'])
+    except jwt.ExpiredSignatureError:
+        raise NotAuthenticated('Unauthenticated')
+
+    task = Task.objects.filter(id=pk, user=payload['id']).first()
 
     serializer = TaskSerializer(task, many=False)
 
@@ -45,6 +71,19 @@ def taskDetail(request, pk):
 
 @api_view(['POST'])
 def taskCreate(request):
+
+    token = request.COOKIES.get('jwt')
+
+    if not token:
+        raise NotAuthenticated('Unauthenticated')
+
+    try:
+        payload = jwt.decode(token, JWT_SECRET, ['HS256'])
+    except jwt.ExpiredSignatureError:
+        raise NotAuthenticated('Unauthenticated')
+
+    request.data['user'] = payload['id']
+
     serializer = TaskSerializer(data=request.data)
 
     if serializer.is_valid():
@@ -56,7 +95,23 @@ def taskCreate(request):
 @api_view(['PATCH'])
 def taskUpdate(request, pk):
 
-    task = Task.objects.get(id=pk)
+    token = request.COOKIES.get('jwt')
+
+    if not token:
+        raise NotAuthenticated('Unauthenticated')
+
+    try:
+        payload = jwt.decode(token, JWT_SECRET, ['HS256'])
+    except jwt.ExpiredSignatureError:
+        raise NotAuthenticated('Unauthenticated')
+
+    task = Task.objects.filter(id=pk, user=payload['id']).first()
+
+    if not task:
+        raise NotFound('Task not found')
+
+    request.data['user'] = payload['id']
+
     serializer = TaskSerializer(instance=task, data=request.data)
 
     if serializer.is_valid():
@@ -67,7 +122,21 @@ def taskUpdate(request, pk):
 
 @api_view(['DELETE'])
 def taskDelete(request, pk):
-    task = Task.objects.get(id=pk)
+
+    token = request.COOKIES.get('jwt')
+
+    if not token:
+        raise NotAuthenticated('Unauthenticated')
+
+    try:
+        payload = jwt.decode(token, JWT_SECRET, ['HS256'])
+    except jwt.ExpiredSignatureError:
+        raise NotAuthenticated('Unauthenticated')
+
+    task = Task.objects.filter(id=pk, user=payload['id']).first()
+
+    if not task:
+        raise NotFound('Task not found!')
 
     task.delete()
